requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
django-taggit==6.1.0
gunicorn==23.0.0
whitenoise==6.6.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
Pillow==11.3.0
email-validator==2.2.0
requests==2.31.0